+++ plugins/source/crowdstrike/codegen/recipes/alerts.go
@@ -1,0 +1,17 @@
+package recipes
+
+import (
+	"github.com/crowdstrike/gofalcon/falcon/models"
+)
+
+func Alerts() []*Resource {
+	resources := []*Resource{
+		{
+			Service:    "alerts",
+			SubService: "query",
+			Struct:     &models.MsaQueryResponse{},
+		},
+	}
+
+	return resources
+}