+++ plugins/source/gcp/resources/services/apikeys/keys_mock_test.go
@@ -8,45 +8,15 @@ import (
 	"github.com/cloudquery/plugin-sdk/faker"
 	"github.com/cloudquery/plugins/source/gcp/client"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-	"net"
 	"testing"
 
-	"cloud.google.com/go/apikeys/apiv2"
-
 	pb "google.golang.org/genproto/googleapis/api/apikeys/v2"
-
-	"google.golang.org/api/option"
 )
 
-func createKeys() (*client.Services, error) {
+func createKeys(gsrv *grpc.Server) error {
 	fakeServer := &fakeKeysServer{}
-	l, err := net.Listen("tcp", "localhost:0")
-	if err != nil {
-		return nil, fmt.Errorf("failed to listen: %w", err)
-	}
-	gsrv := grpc.NewServer()
 	pb.RegisterApiKeysServer(gsrv, fakeServer)
-	fakeServerAddr := l.Addr().String()
-	go func() {
-		if err := gsrv.Serve(l); err != nil {
-			panic(err)
-		}
-	}()
-
-	// Create a client.
-	svc, err := apikeys.NewClient(context.Background(),
-		option.WithEndpoint(fakeServerAddr),
-		option.WithoutAuthentication(),
-		option.WithGRPCDialOption(grpc.WithTransportCredentials(insecure.NewCredentials())),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create grpc client: %w", err)
-	}
-
-	return &client.Services{
-		ApikeysClient: svc,
-	}, nil
+	return nil
 }
 
 type fakeKeysServer struct {
@@ -63,5 +33,5 @@ func (f *fakeKeysServer) ListKeys(context.Context, *pb.ListKeysRequest) (*pb.Lis
 }
 
 func TestKeys(t *testing.T) {
-	client.MockTestHelper(t, Keys(), createKeys, client.TestOptions{})
+	client.MockTestGrpcHelper(t, Keys(), createKeys, client.TestOptions{})
 }