+++ plugins/destination/bigquery/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [1.1.1](https://github.com/cloudquery/cloudquery/compare/plugins-destination-bigquery-v1.1.0...plugins-destination-bigquery-v1.1.1) (2022-12-09)
+
+
+### Bug Fixes
+
+* Fix BigQuery JSON credential loading and allow users to set dataset location in config ([#5527](https://github.com/cloudquery/cloudquery/issues/5527)) ([446e486](https://github.com/cloudquery/cloudquery/commit/446e48648ea7d6e4136e10bec8e260a5efa76c0b))
+
+## [1.1.0](https://github.com/cloudquery/cloudquery/compare/plugins-destination-bigquery-v1.0.1...plugins-destination-bigquery-v1.1.0) (2022-12-08)
+
+
+### Features
+
+* Support service account keys in BigQuery destination ([#5508](https://github.com/cloudquery/cloudquery/issues/5508)) ([5059ecc](https://github.com/cloudquery/cloudquery/commit/5059ecca76b1e186024ac4d582c2ef83f7c12e51))
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.11.1 ([#5458](https://github.com/cloudquery/cloudquery/issues/5458)) ([58b7432](https://github.com/cloudquery/cloudquery/commit/58b74321cd253c9a843c8c103f324abb93952195))
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.11.2 ([#5497](https://github.com/cloudquery/cloudquery/issues/5497)) ([c1876cf](https://github.com/cloudquery/cloudquery/commit/c1876cf793b43d825a25fb3c9ba4996e4b09964f))
+
 ## [1.0.1](https://github.com/cloudquery/cloudquery/compare/plugins-destination-bigquery-v1.0.0...plugins-destination-bigquery-v1.0.1) (2022-12-06)